Fix warnings for iOS7

+++ Project/UISS/UISSUserInterfaceIdiomPreprocessor.m
@@ -22,7 +22,7 @@
     [dictionary enumerateKeysAndObjectsUsingBlock:^(NSString *key, id object, BOOL *stop) {
         UIUserInterfaceIdiom idiom = (UIUserInterfaceIdiom) [self userInterfaceIdiomFromKey:key];
         
-        if (idiom == NSNotFound) {
+        if ((NSInteger)idiom == NSNotFound) {
             [preprocessed setObject:[self preprocessValueIfNecessary:object userInterfaceIdiom:userInterfaceIdiom] forKey:key];
         } else {
             if (idiom == userInterfaceIdiom) {
